feat: timer triggered functions
I created the Azure Functions host framework in the pkg/functions
directory. This framework is responsible for handling the HTTP
communication between the Azure Functions host and the Go function
implementations.

I created the timer package in the pkg directory. This package defines
the Info type that represents the data that the timer trigger sends to
the Azure Function.

I created the timer_trigger example program to demonstrate how to build
a timer-triggered Azure Function.